Clone field using before contributing to class.
Make sure to clone the field before contributing in to the class. This would remove the error where a ManyToManyField's automatically generated "through" model doesn't have the necessary foreign keys for both models linking to the same model. See #5. This commit does not solve the problem when a custom through model is used.